Web or TV Game Assignment
"The purpose of this assignment is to gain some experience with
developing interactive games for
the Web or the TV environment.  You can choose one of these two
platforms you don’t need to do both of them.  In both cases you will be
using the sprite library that we developed in class." - Dr. Mark Green
of UOIT